fix inheritance
